Increased readability of main "join"
Removed a for loop around the select statement used to wait for the producer and consumer goroutines to complete. Instead the select statement just specifies to wait for the other goroutine to complete. This is a reasonable because there are only two threads to wait on, any more and it is probably best to use the previous implementation.